+++ scripts/webframeworks-deploy-tests/hosting/pages/pages/ssg/index.tsx
@@ -1,7 +1,10 @@
+import { useRouter } from "next/router";
+
 export const getStaticProps = async () => {
     return { props: { } };
 }
 
 export default function SSG() {
-    return <>SSG</>;
+    const { locale } = useRouter();
+    return <>SSG { locale }</>;
 }