+++ scripts/webframeworks-deploy-tests/hosting/pages/pages/ssr/index.tsx
@@ -1,7 +1,10 @@
+import { useRouter } from "next/router";
+
 export const getServerSideProps = async () => {
     return { props: { foo: 1 } };
 }
 
 export default function SSR() {
-    return <>SSR</>;
+    const { locale } = useRouter();
+    return <>SSR {locale}</>;
 }