+++ src/extensions/extensionsApi.ts
@@ -3,7 +3,7 @@ import * as clc from "colorette";
 import { marked } from "marked";
 
 import { Client } from "../apiv2";
-import { extensionsOrigin, extensionsPublisherOrigin } from "../api";
+import { extensionsOrigin } from "../api";
 import { FirebaseError } from "../error";
 import { logger } from "../logger";
 import * as operationPoller from "../operation-poller";
@@ -14,20 +14,14 @@ import {
   ExtensionSource,
   ExtensionSpec,
   ExtensionVersion,
-  PublisherProfile,
 } from "./types";
 
 const EXTENSIONS_API_VERSION = "v1beta";
-const PUBLISHER_API_VERSION = "v1beta";
 const PAGE_SIZE_MAX = 100;
 
 const extensionsApiClient = new Client({
   urlPrefix: extensionsOrigin,
   apiVersion: EXTENSIONS_API_VERSION,
-});
-const extensionsPublisherApiClient = new Client({
-  urlPrefix: extensionsPublisherOrigin,
-  apiVersion: PUBLISHER_API_VERSION,
 });
 
 /**
@@ -377,7 +371,7 @@ async function patchInstance(args: {
   return pollRes;
 }
 
-function populateSpec(spec: ExtensionSpec): void {
+export function populateSpec(spec: ExtensionSpec): void {
   if (spec) {
     for (const r of spec.resources) {
       try {
@@ -526,209 +520,6 @@ export async function listExtensionVersions(
 }
 
 /**
- * @param projectId the project for which we are registering a PublisherProfile
- * @param publisherId the desired publisher ID
- */
-export async function getPublisherProfile(
-  projectId: string,
-  publisherId?: string
-): Promise<PublisherProfile> {
-  const res = await extensionsPublisherApiClient.get(`/projects/${projectId}/publisherProfile`, {
-    queryParams:
-      publisherId === undefined
-        ? undefined
-        : {
-            publisherId,
-          },
-  });
-  return res.body as PublisherProfile;
-}
-
-/**
- * @param projectId the project for which we are registering a PublisherProfile
- * @param publisherId the desired publisher ID
- */
-export async function registerPublisherProfile(
-  projectId: string,
-  publisherId: string
-): Promise<PublisherProfile> {
-  const res = await extensionsPublisherApiClient.patch<Partial<PublisherProfile>, PublisherProfile>(
-    `/projects/${projectId}/publisherProfile`,
-    {
-      publisherId,
-      displayName: publisherId,
-    },
-    {
-      queryParams: {
-        updateMask: "publisher_id,display_name",
-      },
-    }
-  );
-  return res.body;
-}
-
-/**
- * @param extensionRef user-friendly identifier for the ExtensionVersion (publisher-id/extension-id@version)
- * @param deprecationMessage the deprecation message
- */
-export async function deprecateExtensionVersion(
-  extensionRef: string,
-  deprecationMessage: string
-): Promise<ExtensionVersion> {
-  const ref = refs.parse(extensionRef);
-  try {
-    const res = await extensionsPublisherApiClient.post<
-      { deprecationMessage: string },
-      ExtensionVersion
-    >(`/${refs.toExtensionVersionName(ref)}:deprecate`, {
-      deprecationMessage,
-    });
-    return res.body;
-  } catch (err: any) {
-    if (err.status === 403) {
-      throw new FirebaseError(
-        `You are not the owner of extension '${clc.bold(
-          extensionRef
-        )}' and don’t have the correct permissions to deprecate this extension version.` + err,
-        { status: err.status }
-      );
-    } else if (err.status === 404) {
-      throw new FirebaseError(`Extension version ${clc.bold(extensionRef)} was not found.`);
-    } else if (err instanceof FirebaseError) {
-      throw err;
-    }
-    throw new FirebaseError(
-      `Error occurred deprecating extension version '${extensionRef}': ${err}`,
-      {
-        status: err.status,
-      }
-    );
-  }
-}
-
-/**
- * @param extensionRef user-friendly identifier for the ExtensionVersion (publisher-id/extension-id@version)
- */
-export async function undeprecateExtensionVersion(extensionRef: string): Promise<ExtensionVersion> {
-  const ref = refs.parse(extensionRef);
-  try {
-    const res = await extensionsPublisherApiClient.post<void, ExtensionVersion>(
-      `/${refs.toExtensionVersionName(ref)}:undeprecate`
-    );
-    return res.body;
-  } catch (err: any) {
-    if (err.status === 403) {
-      throw new FirebaseError(
-        `You are not the owner of extension '${clc.bold(
-          extensionRef
-        )}' and don’t have the correct permissions to undeprecate this extension version.`,
-        { status: err.status }
-      );
-    } else if (err.status === 404) {
-      throw new FirebaseError(`Extension version ${clc.bold(extensionRef)} was not found.`);
-    } else if (err instanceof FirebaseError) {
-      throw err;
-    }
-    throw new FirebaseError(
-      `Error occurred undeprecating extension version '${extensionRef}': ${err}`,
-      {
-        status: err.status,
-      }
-    );
-  }
-}
-
-/**
- * @param extensionVersionRef user-friendly identifier for the extension version (publisher-id/extension-id@1.0.0)
- * @param packageUri public URI of the extension archive (zip or tarball)
- * @param extensionRoot root directory that contains this extension, defaults to "/".
- */
-export async function createExtensionVersionFromLocalSource(args: {
-  extensionVersionRef: string;
-  packageUri: string;
-  extensionRoot?: string;
-}): Promise<ExtensionVersion> {
-  const ref = refs.parse(args.extensionVersionRef);
-  if (!ref.version) {
-    throw new FirebaseError(
-      `Extension version ref "${args.extensionVersionRef}" must supply a version.`
-    );
-  }
-  // TODO(b/185176470): Publishing an extension with a previously deleted name will return 409.
-  // Need to surface a better error, potentially by calling getExtension.
-  const uploadRes = await extensionsPublisherApiClient.post<
-    {
-      versionId: string;
-      extensionRoot: string;
-      remoteArchiveSource: {
-        packageUri: string;
-      };
-    },
-    ExtensionVersion
-  >(`/${refs.toExtensionName(ref)}/versions:createFromSource`, {
-    versionId: ref.version,
-    extensionRoot: args.extensionRoot ?? "/",
-    remoteArchiveSource: {
-      packageUri: args.packageUri,
-    },
-  });
-  const pollRes = await operationPoller.pollOperation<ExtensionVersion>({
-    apiOrigin: extensionsPublisherOrigin,
-    apiVersion: PUBLISHER_API_VERSION,
-    operationResourceName: uploadRes.body.name,
-    masterTimeout: 600000,
-  });
-  return pollRes;
-}
-
-/**
- * @param extensionVersionRef user-friendly identifier for the extension version (publisher-id/extension-id@1.0.0)
- * @param repoUri public GitHub repo URI that contains the extension source
- * @param sourceRef commit hash, branch, or tag to build from the repo
- * @param extensionRoot root directory that contains this extension, defaults to "/".
- */
-export async function createExtensionVersionFromGitHubSource(args: {
-  extensionVersionRef: string;
-  repoUri: string;
-  sourceRef: string;
-  extensionRoot?: string;
-}): Promise<ExtensionVersion> {
-  const ref = refs.parse(args.extensionVersionRef);
-  if (!ref.version) {
-    throw new FirebaseError(
-      `Extension version ref "${args.extensionVersionRef}" must supply a version.`
-    );
-  }
-  // TODO(b/185176470): Publishing an extension with a previously deleted name will return 409.
-  // Need to surface a better error, potentially by calling getExtension.
-  const uploadRes = await extensionsPublisherApiClient.post<
-    {
-      versionId: string;
-      extensionRoot: string;
-      githubRepositorySource: {
-        uri: string;
-        sourceRef: string;
-      };
-    },
-    ExtensionVersion
-  >(`/${refs.toExtensionName(ref)}/versions:createFromSource`, {
-    versionId: ref.version,
-    extensionRoot: args.extensionRoot ?? "/",
-    githubRepositorySource: {
-      uri: args.repoUri,
-      sourceRef: args.sourceRef,
-    },
-  });
-  const pollRes = await operationPoller.pollOperation<ExtensionVersion>({
-    apiOrigin: extensionsPublisherOrigin,
-    apiVersion: PUBLISHER_API_VERSION,
-    operationResourceName: uploadRes.body.name,
-    masterTimeout: 600000,
-  });
-  return pollRes;
-}
-
-/**
  * @param ref user-friendly identifier for the Extension (publisher-id/extension-id)
  * @return the extension
  */
@@ -749,7 +540,7 @@ export async function getExtension(extensionRef: string): Promise<Extension> {
   }
 }
 
-function refNotFoundError(ref: refs.Ref): FirebaseError {
+export function refNotFoundError(ref: refs.Ref): FirebaseError {
   return new FirebaseError(
     `The extension reference '${clc.bold(
       ref.version ? refs.toExtensionVersionRef(ref) : refs.toExtensionRef(ref)